Fri, Jul 28, 2023, 3:49 PM +06:00

+++ beecrowd/1255.c
@@ -19,7 +19,7 @@ int main(void)
         while(text[j] != '\0')
         {
             if((text[j] >= 'A') && (text[j] <= 'Z')) freq[text[j] + 32]++;
-            else freq[(int)text[j]]++;
+            else freq[text[j]]++;
             j++;
         }
